Updated Backend's express-openapi-validator to v4.
