--- new-backend/server/common/oas.js
@@ -1,25 +1,0 @@
-import Express from "express";
-import * as path from "path";
-import errorHandler from "../api/middlewares/error.handler";
-import { OpenApiValidator } from "express-openapi-validator";
-
-export default function oas(app, routes) {
-  const apiSpec = path.join(__dirname, "api.yml");
-  const validateResponses = !!(
-    process.env.OPENAPI_ENABLE_RESPONSE_VALIDATION &&
-    process.env.OPENAPI_ENABLE_RESPONSE_VALIDATION.toLowerCase() === "true"
-  );
-  return new OpenApiValidator({
-    apiSpec,
-    validateResponses,
-    validateRequests: {
-      allowUnknownQueryParameters: true,
-    },
-  })
-    .install(app)
-    .then(() => {
-      app.use(process.env.OPENAPI_SPEC || "/spec", Express.static(apiSpec));
-      routes(app);
-      app.use(errorHandler);
-    });
-}
